authentication/views.py
from drf_yasg.utils import swagger_auto_schema
from rest_framework import generics, permissions, status
from rest_framework.response import Response

from .serializers import RegisterSerializer


class RegisterView(generics.CreateAPIView):
    """
    Registra um novo usuário

    * Cria um novo usuário no banco de dados.
    """

    permission_classes = [permissions.AllowAny]
    serializer_class = RegisterSerializer

    @swagger_auto_schema(
        responses={201: "Usuário criado com sucesso", 400: "Má formatação"},
        request_body=RegisterSerializer,
    )
    def post(self, request):
        user = request.data
        serializer = self.serializer_class(data=user)
        serializer.is_valid(raise_exception=True)
        serializer.save()
        user_data = serializer.data

        return Response(user_data, status.HTTP_201_CREATED)
