pyproject.toml
[tool.poetry]
name = "gallery-django-react"
version = "0.1.0"
description = "Galeria de imagens desenvolvida com Django e React JS."
authors = ["foschieraanderson <foschieraanderson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
drf-yasg = "^1.20.0"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
